+++ frontend/src/metabase/dashboard/containers/DashboardApp/DashboardApp.unit.spec.jsx
@@ -10,12 +10,14 @@ import {
   fireEvent,
 } from "__support__/ui";
 import DashboardApp from "metabase/dashboard/containers/DashboardApp";
+import { BEFORE_UNLOAD_UNSAVED_MESSAGE } from "metabase/hooks/use-before-unload";
 import {
   createMockCard,
   createMockCollection,
   createMockCollectionItem,
   createMockDashboard,
   createMockDatabase,
+  createMockTable,
   createMockUser,
 } from "metabase-types/api/mocks";
 import { createMockDashboardState } from "metabase-types/store/mocks";
@@ -26,32 +28,26 @@ import {
   setupDashboardEndpoints,
   setupDatabasesEndpoints,
   setupSearchEndpoints,
+  setupTableEndpoints,
 } from "__support__/server-mocks";
 import { createMockEntitiesState } from "__support__/store";
+import { callMockEvent } from "__support__/events";
 
-const TEST_DASHBOARD = createMockDashboard({
-  id: 1,
-  name: "Example",
-});
-
-const TEST_COLLECTION = createMockCollection({
-  id: "root",
-});
+const TEST_DASHBOARD = createMockDashboard();
+const TEST_COLLECTION = createMockCollection();
 
 const TEST_DATABASE_WITH_ACTIONS = createMockDatabase({
-  id: 1,
-  name: "Test Database",
   settings: { "database-enable-actions": true },
 });
 
 const TEST_COLLECTION_ITEM = createMockCollectionItem({
   collection: TEST_COLLECTION,
-  id: 1,
-  name: "Test search response",
   model: "dataset",
 });
 
 const TEST_CARD = createMockCard();
+
+const TEST_TABLE = createMockTable();
 
 async function setup({ user = createMockUser() }) {
   setupDatabasesEndpoints([TEST_DATABASE_WITH_ACTIONS]);
@@ -59,13 +55,14 @@ async function setup({ user = createMockUser() }) {
   setupCollectionsEndpoints([]);
   setupCollectionItemsEndpoint(TEST_COLLECTION);
   setupSearchEndpoints([TEST_COLLECTION_ITEM]);
-
   setupCardsEndpoints([TEST_CARD]);
+  setupTableEndpoints([TEST_TABLE]);
 
   fetchMock.get("path:/api/bookmark", []);
   fetchMock.get("path:/api/action", []);
 
   window.HTMLElement.prototype.scrollIntoView = function () {};
+  const mockEventListener = jest.spyOn(window, "addEventListener");
 
   const DashboardAppContainer = props => {
     return (
@@ -95,7 +92,7 @@ async function setup({ user = createMockUser() }) {
     screen.queryAllByTestId("loading-spinner"),
   );
 
-  return { container };
+  return { container, mockEventListener };
 }
 
 const navigateToDashboardActionsEditor = async () => {
@@ -125,10 +122,40 @@ describe("DashboardApp", function () {
     jest.clearAllMocks();
   });
 
+  describe("beforeunload events", () => {
+    afterEach(() => {
+      jest.clearAllMocks();
+    });
+
+    it("should have a beforeunload event when the user tries to leave a dirty dashboard", async function () {
+      const { mockEventListener } = await setup({});
+
+      userEvent.click(screen.getByLabelText("Edit dashboard"));
+      userEvent.click(screen.getByTestId("dashboard-name-heading"));
+      userEvent.type(screen.getByTestId("dashboard-name-heading"), "a");
+      // need to click away from the input to trigger the isDirty flag
+      userEvent.tab();
+
+      const mockEvent = callMockEvent(mockEventListener, "beforeunload");
+
+      expect(mockEvent.preventDefault).toHaveBeenCalled();
+      expect(mockEvent.returnValue).toEqual(BEFORE_UNLOAD_UNSAVED_MESSAGE);
+    });
+
+    it("should not have a beforeunload event when the dashboard is unedited", async function () {
+      const { mockEventListener } = await setup({});
+
+      userEvent.click(screen.getByLabelText("Edit dashboard"));
+
+      const mockEvent = callMockEvent(mockEventListener, "beforeunload");
+      expect(mockEvent.preventDefault).not.toHaveBeenCalled();
+      expect(mockEvent.returnValue).toBe(undefined);
+    });
+  });
+
   describe("ActionCreatorModal onClickOutside behavior", () => {
     it("should not close ActionCreator modal when clicking outside modal", async () => {
       const { container } = await setup({});
-
       await navigateToDashboardActionsEditor();
 
       fireEvent.click(container.ownerDocument.body);
@@ -138,10 +165,13 @@ describe("DashboardApp", function () {
     });
     it("should close ActionCreator modal when clicking modal's 'Cancel' button", async () => {
       await setup({});
-
       await navigateToDashboardActionsEditor();
 
-      userEvent.click(screen.getByText("Cancel"));
+      userEvent.click(
+        within(screen.getByTestId("action-creator-modal-actions")).getByText(
+          "Cancel",
+        ),
+      );
 
       expect(
         screen.queryByTestId("mock-native-query-editor"),
